+++ scripts/changelog.ts
@@ -898,6 +898,22 @@ async function exec(): Promise<void> {
       date: '2021-07-01',
       fixed: ['Fix update of cache in mutation "setSubscriptions"'],
     },
+    {
+      tagName: 'v0.26.2',
+      date: '2021-07-07',
+      added: ['Add property User.activityByType', 'config hi.serlo.org/serlo'],
+      breakingChanges: [
+        'Add upper limit of returns for all connections (default value = 500)',
+      ],
+    },
+    {
+      tagName: 'v0.26.3',
+      date: '2021-07-08',
+      added: [
+        'Make Solution.revisions not optional',
+        'Increase upper limit of events query to 500',
+      ],
+    },
   ])
 
   await writeFile(path.join(__dirname, '..', 'CHANGELOG.md'), content)